+++ CHANGELOG.md
@@ -2,13 +2,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## v2.0.0 - 2019-06-11
+
+### Changed
+ - QueryExecutor now returns statements as a Statement interface object rather than a raw string
+
 ## v1.2.0 - 2015-12-22
 
 ### Added
  - Implemented `ORDER BY` for read queries
 
 ### Changed
-- Updated mock tables to use new method of decoding, now also supports embedded and non-key map types.
+ - Updated mock tables to use new method of decoding, now also supports embedded and non-key map types.
 
 ## v1.1.0 - 2015-11-27